+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog for AWS SDK for JavaScript
-<!--LATEST=2.1259.0-->
+<!--LATEST=2.1260.0-->
 <!--ENTRYINSERT-->
+
+## 2.1260.0
+* feature: Appflow: Adding support for Amazon AppFlow to transfer the data to Amazon Redshift databases through Amazon Redshift Data API service. This feature will support the Redshift destination connector on both public and private accessible Amazon Redshift Clusters and Amazon Redshift Serverless.
+* feature: KinesisAnalyticsV2: Support for Apache Flink 1.15 in Kinesis Data Analytics.
 
 ## 2.1259.0
 * feature: Route53: Amazon Route 53 now supports the Asia Pacific (Hyderabad) Region (ap-south-2) for latency records, geoproximity records, and private DNS for Amazon VPCs in that region.